symlink readme to root
I am not completely sure that having the segmentation tool in the same
repo with the web app is the best, but at the moment I am doing my
best to not overthink things. I don't want to have to write a separate
read me so this will do for now.